@@ -7,8 +7,8 @@ package mock_shell
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	logrus "github.com/sirupsen/logrus"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockInterface is a mock of Interface interface.
